# docs/swagger.yaml
basePath: /
definitions:
...
  input.CreateEventInput:
    description: Dados necessários para criar um novo evento no sistema
    properties:
      cityId:
        description: ID da cidade onde o evento ocorrerá
        example: 1
        type: integer
      date:
        description: Data e hora do evento no formato RFC3339 (2006-01-02T15:04:05Z07:00)
        example: "2024-12-25T20:00:00-03:00"
        type: string
      description:
        description: Descrição do evento
        example: Passeio ciclístico noturno pela cidade
        type: string
      groupId:
        description: ID do grupo associado ao evento (opcional)
        example: 5
        type: integer
      location:
        allOf:
        - $ref: '#/definitions/entities.Coordinates'
        description: Coordenadas geográficas da localização do evento
      name:
        description: Nome do evento
        example: Pedal Noturno
        type: string
      photoContentType:
        description: 'Tipo de conteúdo da foto do evento (opcional: image/jpeg, image/png,
          image/webp)'
        enum:
        - image/jpeg
        - image/png
        - image/webp
        example: image/jpeg
        type: string
      visibility:
        description: Visibilidade do evento (public ou private)
        enum:
        - public
        - private
        example: public
        type: string
    required:
    - cityId
    - date
    - description
    - location
    - name
    - visibility
    type: object
...
  input.PendingInviteAnswer:
    description: Resposta para aceitar ou rejeitar um convite pendente para participar
      de um evento
    properties:
      answer:
        description: Resposta ao convite (accept ou reject)
        enum:
        - accept
        - reject
        example: accept
        type: string
    required:
    - answer
    type: object
...
  input.UpdateEventInput:
    description: Dados opcionais para atualizar informações de um evento existente
    properties:
      cityId:
        description: ID da cidade (opcional, requer Location se fornecido)
        example: 2
        type: integer
      date:
        description: 'Data e hora do evento no formato RFC3339 (opcional: 2006-01-02T15:04:05Z07:00)'
        example: "2024-12-26T20:00:00-03:00"
        type: string
      description:
        description: Descrição do evento (opcional)
        example: Nova descrição do evento
        type: string
      groupId:
        description: ID do grupo associado (opcional)
        example: 6
        type: integer
      location:
        allOf:
        - $ref: '#/definitions/entities.Coordinates'
        description: Coordenadas geográficas da localização (opcional, requer CityID
          se fornecido)
      name:
        description: Nome do evento (opcional)
        example: Novo Nome do Evento
        type: string
      visibility:
        description: 'Visibilidade do evento (opcional: public ou private)'
        enum:
        - public
        - private
        example: private
        type: string
    type: object
...
  response.GetEventResponse:
    description: Informações completas de um evento retornado na busca/listagem
    properties:
      address:
        description: Endereço do evento (formato JSON)
        type: object
      coordinates:
        description: Coordenadas geográficas do evento (formato JSON)
        type: object
      date:
        description: Data e hora do evento
        example: "2024-12-25T20:00:00Z"
        type: string
      description:
        description: Descrição do evento
        example: Passeio ciclístico noturno pela cidade
        type: string
      group:
        description: Informações do grupo associado (opcional, formato JSON)
        type: object
      id:
        description: ID do evento
        example: 456
        type: integer
      memberRole:
        description: Papel do usuário no evento (owner, admin, member, subscriber)
        example: subscriber
        type: string
      name:
        description: Nome do evento
        example: Pedal Noturno
        type: string
      photo:
        description: URL da foto do evento (opcional)
        example: https://example.com/events/456/photo.jpg
        type: string
      subscribersCount:
        description: Número de inscritos no evento (opcional)
        example: 25
        type: integer
      visibility:
        description: Visibilidade do evento (public ou private)
        example: public
        type: string
    type: object
...
  response.PresginedEventPhotoResponse:
    description: URL pré-assinada e chave para upload de foto do evento
    properties:
      eventId:
        description: ID do evento
        example: 456
        type: integer
      key:
        description: Chave da foto no storage
        example: events/456/photo.jpg
        type: string
      url:
        description: URL pré-assinada para upload da foto
        example: https://s3.amazonaws.com/bucket/events/456/photo.jpg?X-Amz-Signature=...
        type: string
    type: object
...
  src_core_event_controller.ConfirmNewPhotoRequest:
    properties:
      key:
        type: string
    required:
    - key
    type: object
  src_core_event_controller.ErrorMessageResponse:
    properties:
      message:
        type: string
    type: object
  src_core_event_controller.ValidationErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: API do Revit integrada com Swagger via Fiber.
  title: Revit API
  version: "1.0"
paths:
...
  /api/event:
    get:
      consumes:
      - application/json
      description: Retorna a lista de eventos do usuário autenticado
      parameters:
      - description: ID da cidade para filtrar (opcional)
        example: 1
        in: query
        name: cityID
        type: integer
      - description: 'Data inicial para filtrar eventos (opcional, formato: 2006-01-02)'
        example: "2024-12-01"
        in: query
        name: fromDate
        type: string
      - description: Latitude para filtrar eventos próximos (opcional, requer Longitude)
        example: -23.5505
        in: query
        name: latitude
        type: number
      - description: Limite de resultados por página (opcional)
        example: 10
        in: query
        name: limit
        type: integer
      - description: Longitude para filtrar eventos próximos (opcional, requer Latitude)
        example: -46.6333
        in: query
        name: longitude
        type: number
      - description: 'Tipo de membro do usuário no evento (opcional: owner, admin,
          member)'
        enum:
        - owner
        - admin
        - member
        example: member
        in: query
        name: memberType
        type: string
      - description: Nome do evento para busca (opcional)
        example: Pedal
        in: query
        name: name
        type: string
      - description: Número da página (opcional, deve ser maior que 0)
        example: 1
        in: query
        name: page
        type: integer
      - description: 'Data final para filtrar eventos (opcional, formato: 2006-01-02)'
        example: "2024-12-31"
        in: query
        name: toDate
        type: string
      - description: 'Visibilidade do evento (opcional: public ou private)'
        enum:
        - public
        - private
        example: public
        in: query
        name: visibility
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de eventos
          schema:
            items:
              $ref: '#/definitions/response.GetEventResponse'
            type: array
        "400":
          description: Erro na validação dos parâmetros
          schema:
            $ref: '#/definitions/src_core_event_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Listar eventos
      tags:
      - Events
    post:
      consumes:
      - application/json
      description: Cria um novo evento
      parameters:
      - description: Dados do evento
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/input.CreateEventInput'
      produces:
      - application/json
      responses:
        "200":
          description: Evento criado com sucesso
          schema:
            $ref: '#/definitions/response.PresginedEventPhotoResponse'
        "400":
          description: Erro na validação dos dados
          schema:
            $ref: '#/definitions/src_core_event_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Criar evento
      tags:
      - Events
  /api/event/{eventId}:
    patch:
      consumes:
      - application/json
      description: Atualiza os dados de um evento
      parameters:
      - description: ID do evento
        in: path
        name: eventId
        required: true
        type: integer
      - description: Dados do evento para atualização
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/input.UpdateEventInput'
      produces:
      - application/json
      responses:
        "204":
          description: Evento atualizado com sucesso
        "400":
          description: Erro na validação dos dados
          schema:
            $ref: '#/definitions/src_core_event_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "403":
          description: Sem permissão para atualizar o evento
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "404":
          description: Evento não encontrado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Atualizar evento
      tags:
      - Events
  /api/event/{eventId}/invite:
    patch:
      consumes:
      - application/json
      description: Aceita ou rejeita um convite de evento pendente
      parameters:
      - description: ID do evento
        in: path
        name: eventId
        required: true
        type: integer
      - description: Resposta ao convite (accept/reject)
        in: body
        name: answer
        required: true
        schema:
          $ref: '#/definitions/input.PendingInviteAnswer'
      produces:
      - application/json
      responses:
        "204":
          description: Resposta processada com sucesso
        "400":
          description: Erro na validação dos dados
          schema:
            $ref: '#/definitions/src_core_event_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "404":
          description: Convite não encontrado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Responder convite pendente
      tags:
      - Events
  /api/event/{eventId}/invite/{invitedId}:
    post:
      consumes:
      - application/json
      description: Envia um convite para um usuário participar de um evento
      parameters:
      - description: ID do evento
        in: path
        name: eventId
        required: true
        type: integer
      - description: ID do usuário a ser convidado
        in: path
        name: invitedId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Convite enviado com sucesso
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "403":
          description: Sem permissão para convidar
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "404":
          description: Evento ou usuário não encontrado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Convidar usuário para evento
      tags:
      - Events
  /api/event/{eventId}/subscriber:
    delete:
      consumes:
      - application/json
      description: Remove o usuário autenticado de um evento
      parameters:
      - description: ID do evento
        in: path
        name: eventId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Inscrição cancelada com sucesso
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "404":
          description: Evento ou inscrição não encontrada
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Cancelar inscrição em evento
      tags:
      - Events
    post:
      consumes:
      - application/json
      description: Adiciona o usuário autenticado como participante de um evento
      parameters:
      - description: ID do evento
        in: path
        name: eventId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Usuário inscrito no evento
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "404":
          description: Evento não encontrado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Inscrever-se em evento
      tags:
      - Events
  /api/event/{eventId}/subscriber/{subscriberId}:
    delete:
      consumes:
      - application/json
      description: Remove um participante de um evento (apenas administradores)
      parameters:
      - description: ID do evento
        in: path
        name: eventId
        required: true
        type: integer
      - description: ID do participante a ser removido
        in: path
        name: subscriberId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Participante removido com sucesso
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "403":
          description: Sem permissão para remover participante
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "404":
          description: Evento ou participante não encontrado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Remover participante do evento
      tags:
      - Events
  /api/event/invite:
    get:
      consumes:
      - application/json
      description: Retorna os convites de eventos pendentes do usuário autenticado
      parameters:
      - description: Limite de resultados por página (opcional, máximo 20)
        example: 10
        in: query
        maximum: 20
        name: limit
        type: integer
      - description: Número da página (opcional, deve ser maior que 0)
        example: 1
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de convites pendentes
          schema:
            items:
              $ref: '#/definitions/response.GetPendingInvitesResponse'
            type: array
        "400":
          description: Erro na validação dos parâmetros
          schema:
            $ref: '#/definitions/src_core_event_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Listar convites pendentes
      tags:
      - Events
  /api/event/photo/{eventId}:
    patch:
      consumes:
      - application/json
      description: Confirma o upload de foto após upload na URL pré-assinada
      parameters:
      - description: ID do evento
        in: path
        name: eventId
        required: true
        type: integer
      - description: Confirmação do upload
        in: body
        name: confirmation
        required: true
        schema:
          $ref: '#/definitions/src_core_event_controller.ConfirmNewPhotoRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Foto confirmada com sucesso
        "400":
          description: Erro na validação dos dados
          schema:
            $ref: '#/definitions/src_core_event_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "403":
          description: Sem permissão para atualizar o evento
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "404":
          description: Evento não encontrado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Confirmar upload de foto do evento
      tags:
      - Events
    post:
      consumes:
      - application/json
      description: Solicita uma URL pré-assinada para upload de foto do evento
      parameters:
      - description: ID do evento
        in: path
        name: eventId
        required: true
        type: integer
      - description: Dados da requisição de foto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/input.RequestNewPhotoInput'
      produces:
      - application/json
      responses:
        "201":
          description: URL pré-assinada para upload
          schema:
            $ref: '#/definitions/response.PresginedEventPhotoResponse'
        "400":
          description: Erro na validação dos dados
          schema:
            $ref: '#/definitions/src_core_event_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "403":
          description: Sem permissão para atualizar o evento
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
        "404":
          description: Evento não encontrado
          schema:
            $ref: '#/definitions/src_core_event_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Solicitar upload de foto do evento
      tags:
      - Events
...
securityDefinitions:
  BearerAuth:
    description: 'Digite "Bearer " seguido do token JWT. Exemplo: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"